requirements.txt
python-dotenv== 1.0.0 
aiogram == 3.2.0
psycopg==3.1.14 
psycopg-binary==3.1.14